requirements.txt
pybit>=5.13.0
websockets>=12.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0